+++ pkgs/cargoHelperFunctionsHook.nix
@@ -1,0 +1,7 @@
+{ makeSetupHook
+}:
+
+makeSetupHook
+{
+  name = "cargoHelperFunctionsHook";
+} ./cargoHelperFunctions.sh